[AD-512] Revert UserSecret-related changes (#46)
Problem: we've made some changes related to UserSecret because we wanted to
use this type in Ariadne, but not exactly this type as is. However, we don't
use this type anymore. And we want to keep diff with IOHK version as small as
possible to make merging IOHK version easier in future.
Solution: revert all changes, so that files in question correspond to upstream
cardano-sl-1.3.2.